# test/swagger.json
{
    "schemes": [],
    "swagger": "2.0",
    "info": {
        "description": "",
        "title": "Idena flip words translation API",
        "contact": {},
        "license": {
            "name": "Apache 2.0"
        },
        "version": "0.1.0"
    },
    "host": "localhost:82",
    "basePath": "/",
    "paths": {
        "/translation": {
            "post": {
                "tags": [
                    "Translation"
                ],
                "summary": "Create or update translation",
                "operationId": "submitTranslation",
                "parameters": [
                    {
                        "description": "translation details",
                        "name": "translation",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/SubmitTranslationRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/SubmitTranslationResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
        "/vote": {
            "post": {
                "tags": [
                    "Translation"
                ],
                "summary": "Vote for or against translation",
                "operationId": "vote",
                "parameters": [
                    {
                        "description": "vote details",
                        "name": "vote",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/VoteRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/VoteResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
        "/word/{word}/language/{language}/confirmed-translation": {
            "get": {
                "tags": [
                    "Translation"
                ],
                "summary": "Get confirmed translation",
                "operationId": "getConfirmedTranslation",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "word id",
                        "name": "word",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "language",
                        "name": "language",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/GetConfirmedTranslationResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
        "/word/{word}/language/{language}/translations": {
            "get": {
                "tags": [
                    "Translation"
                ],
                "summary": "Get translations sorted by rating",
                "operationId": "getTranslations",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "word id",
                        "name": "word",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "language",
                        "name": "language",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "continuation token to get next translations",
                        "name": "continuation-token",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/GetTranslationsResponse"
                        },
                        "headers": {
                            "continuation-token": {
                                "type": "string",
                                "description": "continuation token"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "GetConfirmedTranslationResponse": {
            "type": "object",
            "properties": {
                "translation": {
                    "type": "object",
                    "$ref": "#/definitions/Translation"
                }
            }
        },
        "GetTranslationsResponse": {
            "type": "object",
            "properties": {
                "translations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Translation"
                    }
                }
            }
        },
        "SubmitTranslationRequest": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string",
                    "maxLength": 100
                },
                "language": {
                    "type": "string",
                    "example": "en"
                },
                "name": {
                    "type": "string",
                    "maxLength": 20
                },
                "signature": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string",
                    "example": "2020-01-01T00:00:00Z"
                },
                "word": {
                    "type": "integer",
                    "maximum": 3299
                }
            }
        },
        "SubmitTranslationResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "resCode": {
                    "type": "integer",
                    "enum": [
                        0,
                        1,
                        2,
                        4
                    ]
                },
                "translationId": {
                    "type": "string"
                }
            }
        },
        "Translation": {
            "type": "object",
            "properties": {
                "confirmed": {
                    "type": "boolean"
                },
                "description": {
                    "type": "string"
                },
                "downVotes": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "upVotes": {
                    "type": "integer"
                }
            }
        },
        "VoteRequest": {
            "type": "object",
            "properties": {
                "signature": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string",
                    "example": "2020-01-01T00:00:00Z"
                },
                "translationId": {
                    "type": "string"
                },
                "up": {
                    "type": "boolean"
                }
            }
        },
        "VoteResponse": {
            "type": "object",
            "properties": {
                "downVotes": {
                    "type": "integer"
                },
                "error": {
                    "type": "string"
                },
                "resCode": {
                    "type": "integer",
                    "enum": [
                        0,
                        3,
                        4,
                        5
                    ]
                },
                "upVotes": {
                    "type": "integer"
                }
            }
        }
    }
}